add pre_gen hook with basic command checking
This is related to issue #1. The pre_gen_project module is
executed by cookiecutter before a project is generated.

Currently we're using a helper function that tries to determine
if a given command is available. This function, command_exists,
aims to be compatible with both Python 2 and 3 and among OSes.