Enable autoformat on save for VSCode
Honestly, I don't like to have this setting globally enabled but
the Black autoformater plays a critical role on this cookiecutter.
In fact, the pre-commit hook will BREAK if there are unformatted
files, prompting the user to re-add them.

So it makes sense to me that the editor autoformats on save.